+++ TOTID/Assets/Scripts/Map/ScGps.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fd47936af2ef3934cbf42049dea33388
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
